style stake overview (summary & elected)
