+++ pages/stake.tsx
@@ -11,10 +11,10 @@ import { NextSeo } from "next-seo";
 import StakeForm from "@/components/StakeForm";
 import StakeAmountInput from "@/components/StakeAmountInput";
 import StakeActionsPair from "@/components/StakeActionsPair";
-// import StakeElectedInfo from "@/components/StakeElectedInfo";
+import StakeElected from "@/components/StakeElected";
 import StakeNominate from "@/components/StakeNominate";
 import StakeOverview from "@/components/StakeOverview";
-import StakeElected from "@/components/StakeElected";
+import StakeSummary from "@/components/StakeSummary";
 
 export async function getStaticProps() {
 	const api = await Api.create({ provider: API_URL });
@@ -32,16 +32,16 @@ const Pool: VFC<{ stakeAssets: StakeAssets }> = ({ stakeAssets }) => {
 		<StakeProvider stakeAssets={stakeAssets}>
 			<div css={styles.root}>
 				<NextSeo title="CENNZnet Staking" />
-				<MainPanel defaultTitle="CENNZnet Staking">
+				<MainPanel defaultTitle="Staking">
 					<StakeForm>
 						<StakeActionsPair />
 						<StakeAmountInput />
 						<StakeNominate />
-						{/*<StakeElectedInfo />*/}
 					</StakeForm>
 				</MainPanel>
-				<MainPanel defaultTitle="Staking Overview">
+				<MainPanel defaultTitle="Overview">
 					<StakeOverview>
+						<StakeSummary />
 						<StakeElected />
 					</StakeOverview>
 				</MainPanel>
